+++ src/cast_core/src/actors/MultiCollectionUpdate.cpp
@@ -29,14 +29,14 @@ struct MultiCollectionUpdate::PhaseConfig {
           minDelay{context.get<TimeSpec, false>("MinDelay").value_or(TimeSpec(0))} {}
 
     mongocxx::database database;
-    uint64_t numCollections;
+    size_t numCollections;
     std::unique_ptr<value_generators::DocumentGenerator> queryDocument;
     std::unique_ptr<value_generators::DocumentGenerator> updateDocument;
     // TODO: Enable passing in update options.
     //    std::unique_ptr<value_generators::DocumentGenerator>  updateOptions;
 
     // uniform distribution random int for selecting collection
-    std::uniform_int_distribution<uint64_t> uniformDistribution;
+    std::uniform_int_distribution<size_t> uniformDistribution;
     std::chrono::milliseconds minDelay;
 };
